changed interrupt index to be 64 bits long and fixed a bad include in adaptec_ctrl.hh
cpu/base_cpu.cc:
    changed index to 64bits
cpu/base_cpu.hh:
    changed index to 64 bits

--HG--
extra : convert_revision : e70d5f09f6066b90fca82cae22bb7d7eb705d65e

+++ cpu/base_cpu.cc
@@ -196,7 +196,7 @@ BaseCPU::post_interrupt(int int_num, int index)
     if (int_num < 0 || int_num >= NumInterruptLevels)
         panic("int_num out of bounds\n");
 
-    if (index < 0 || index >= sizeof(uint8_t) * 8)
+    if (index < 0 || index >= sizeof(uint64_t) * 8)
         panic("int_num out of bounds\n");
 
     AlphaISA::check_interrupts = 1;
@@ -212,7 +212,7 @@ BaseCPU::clear_interrupt(int int_num, int index)
     if (int_num < 0 || int_num >= NumInterruptLevels)
         panic("int_num out of bounds\n");
 
-    if (index < 0 || index >= sizeof(uint8_t) * 8)
+    if (index < 0 || index >= sizeof(uint64_t) * 8)
         panic("int_num out of bounds\n");
 
     interrupts[int_num] &= ~(1 << index);
